refactor: move truncate_k8s_resource_name to helpersmanager
