+++ gateway-api-integrator/src/helpers.py
@@ -1,0 +1,32 @@
+# Copyright 2026 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+"""Helper methods."""
+
+import hashlib
+
+_K8S_RESOURCE_NAME_MAX_LENGTH = 63
+
+
+def truncate_k8s_resource_name(name: str) -> str:
+    """Truncate a Kubernetes resource name to fit within the 63-character limit.
+
+    If the name exceeds 63 characters, it is truncated and a short hash suffix
+    is appended to ensure uniqueness. The resulting name is at most 63
+    characters long.
+
+    Args:
+        name: The desired resource name.
+
+    Returns:
+        The name, possibly truncated with a hash suffix appended.
+    """
+    if len(name) <= _K8S_RESOURCE_NAME_MAX_LENGTH:
+        return name
+    # Use an 8-char hex digest for collision avoidance. usedforsecurity=False
+    # marks this as a non-security use, which also allows MD5 in FIPS environments.
+    suffix = hashlib.md5(name.encode(), usedforsecurity=False).hexdigest()[:8]
+    # Truncate the name leaving room for a dash and the 8-char suffix.
+    max_prefix_length = _K8S_RESOURCE_NAME_MAX_LENGTH - len(suffix) - 1
+    truncated = name[:max_prefix_length].rstrip("-")
+    return f"{truncated}-{suffix}"
